Laberinto

Documentation

data Laberinto

Tipo de datos Laberinto.

Constructors

Trifurcacion (Maybe Laberinto) (Maybe Laberinto) (Maybe Laberinto)
Constructor de una trifurcación de un laberinto en maybe laberintos con caminos en el orden izquierda, recto y derecha.

Tesoro String (Maybe Laberinto)
Constructor de un tesoro en el camino de un laberinto.

caminoSinSalida

Arguments

:: Laberinto
Valor de retorno: Trifurcacion donde todos los caminos conducen a Nothing.

Función que retorna un camino sin salida.

crearTesoro

Arguments

:: [Char]
String que posee la descripción de un Tesoro.

-> Laberinto
Laberinto a agregar en el camino del tesoro.

-> Laberinto
Valor de retorno: Tesoro.

Función que crea un Tesoro.

agregarLaberinto

Arguments

:: Laberinto
Trifurcación base donde se agregará un nuevo Laberinto.

-> Laberinto
Laberinto que se agregará a una Trifurcación.

-> Char
Indicador de cuál camino los relaciona.

-> Laberinto
Valor de retorno: Trifurcación en la cual el Laberinto fue agregado en la posición del camino dado por el indicador.

Función que agrega un Laberinto en el camino de una Trifurcación, dicho camino es determinado mediante un indicador. El indicador consiste en un char i, r o d que indica si se agrega el Laberinto a la izquierda, recto o a la derecha respectivamente.

voltearIzquierda

Arguments

:: Laberinto
Laberinto a recorrer.

-> Laberinto
Valor de retorno: Laberinto que comienza al voltear al camino de la izquierda.

Función que voltea al camino de la izquierda de un Laberinto dado.

voltearDerecha

Arguments

:: Laberinto
Laberinto a recorrer.

-> Laberinto
Valor de retorno: Laberinto que comienza al voltear al camino de la derecha.

Función que voltea al camino de la derecha de un Laberinto dado.

irRecto

Arguments

:: Laberinto
Laberinto a recorrer.

-> Laberinto
Valor de retorno: Laberinto que comienza al seguir por el camino recto.

Función que continua por el camino recto de un Laberinto dado.

recorrerLaberinto

Arguments

:: Laberinto
Laberinto a recorrer.

-> [Char]
String que contiene una ruta especificada con indicadores.

-> Laberinto
Valor de retorno: Laberinto que comienza en el punto al que conduce esa ruta.

Función que recorre una ruta en un Laberinto dado. La ruta se especifica con indicadores que vienen dados por char i, r o d si es izquierda, recto o derecha respectivamente (por ejemplo dird representa el camino derecha-izquierda-recto-derecha).

construirLaberinto

Arguments

:: Laberinto
Laberinto a agregar en la primera dirección de la ruta.

-> [Char]
String que contiene una ruta especificada con indicadores.

-> Laberinto
Valor de retorno: Trifurcación con el Laberinto parámetro agregado en la primera dirección de la ruta.

Función que construye un Laberinto a partir de una ruta y otro Laberinto, el cual es agregado en la primera dirección que se especifica en la ruta.

evaluarLaberinto

Arguments

:: Laberinto
Laberinto a evaluar.

-> Char
Valor de retorno: Caracter que indica qué tipo de Trifurcación es.

Función que evalúa si una Trifurcación es normal, un camino sin salida o una habitación con un Tesoro.

getTesoroStr

Arguments

:: Laberinto
Tesoro del cual se desea saber la descripción.

-> String
Valor de retorno: String que contiene la descripción del Tesoro.

Función que retorna la descripción de un Tesoro.

reportarParedAbierta

Arguments

:: Laberinto
Laberinto a recorrer.

-> [Char]
Ruta a seguir.

-> Laberinto
Valor de retorno:

Funcion que recibe un laberinto y un string, y recorre el laberinto hasta acabar el string o hasta llegar a una pared, donde construye el laberinto con el resto del string y lo agrega en la direccion correspondiente a la cabeza del string restante. Más detalles en README.txt.

reportarDerrumbe

Arguments

:: Laberinto
Laberinto a recorrer.

-> [Char]
Ruta a seguir.

-> [Char]
String cuyo primer caracter es la dirección a derrumbar.

-> Laberinto
Valor de retorno:

Funcion que recibe un laberinto y dos string, y recorre el laberinto hasta acabar el string donde toma la cabeza del segundo string (un char) para reemplazar el laberinto en esa direccion por Nothing. Más detalles en README.txt.

escribirLaberinto

Arguments

:: FilePath
Ruta del archivo a escribir.

-> Laberinto
Laberinto que se desea escribir.

-> IO ()
Valor de retorno: IO.

Función auxiliar que escribe un Laberinto en un archivo.