Main

Documentation

main

Arguments

:: IO ()
Valor de retorno: IO.

Main

laberintoEnMem

Arguments

:: Laberinto
Valor de retorno: Laberinto vacío.

Función que crea una instancia de un Laberinto vacío en memoria.

opcionesDisponibles

Arguments

:: [(String, String)]
Valor de retorno: Diccionario de número de opción y descripción de la opción.

Función que define las opciones del menú.

printMenu

Arguments

:: IO ()
Valor de retorno: IO.

Función para mostrar el menú en pantalla.

instruccionesRuta

Arguments

:: IO ()
Valor de retorno: IO.

Función para mostrar en pantalla las instrucciones de cómo recibir una ruta para un Laberinto.

comenzarLaberintoNuevo

Arguments

:: IO ()
Valor de retorno: IO.

Función para construir un nuevo Laberinto a partir de una ruta dada por el usuario.

construirLaberinto

Arguments

:: Laberinto
Laberinto a agregar en la primera dirección de la ruta.

-> [Char]
String que contiene una ruta especificada con indicadores.

-> Laberinto
Valor de retorno: Trifurcación con el Laberinto parámetro agregado en la primera dirección de la ruta.

Función que construye un Laberinto a partir de una ruta y otro Laberinto, el cual es agregado en la primera dirección que se especifica en la ruta.

mostrarRecibirOpciones

Arguments

:: Laberinto
Laberinto que se tiene en memoria.

-> IO ()
Valor de retorno: IO.

Función para mostrar las opciones y recibir las introducidas por el usuario.