Laberinto

Documentation

data Laberinto

Tipo de datos Laberinto.

Constructors

Trifurcacion (Maybe Laberinto) (Maybe Laberinto) (Maybe Laberinto)
Constructor de una trifurcación de un laberinto en maybe laberintos con caminos en el orden izquierda, recto y derecha.

Tesoro String (Maybe Laberinto)
Constructor de un tesoro en el camino de un laberinto.

caminoSinSalida :: Laberinto

Función que retorna un camino sin salida, es decir una Trifurcacion donde todos los caminos conducen a Nothing.

agregarLaberinto :: Laberinto -> Laberinto -> Char -> Laberinto

Función que recibe una Trifurcacion, un Laberinto y un indicador de cual camino los relaciona y retorna una Trifurcacion en la que se indica que dicho camino especificado en el indicador conduce al Laberinto dado como parámetro. El indicador viene dado por un char i, r o d que indica si se agrega el Laberinto a la izquierda, recto o a la derecha respectivamente.