Laberinto

Documentation

data Laberinto

Tipo de datos Laberinto.

Constructors

Trifurcacion (Maybe Laberinto) (Maybe Laberinto) (Maybe Laberinto)
Constructor de una trifurcación de un laberinto en maybe laberintos con caminos en el orden izquierda, recto y derecha.

Tesoro String (Maybe Laberinto)
Constructor de un tesoro en el camino de un laberinto.

caminoSinSalida :: Laberinto

Función que retorna un camino sin salida, es decir una Trifurcacion donde todos los caminos conducen a Nothing.

agregarLaberinto :: Laberinto -> Laberinto -> Char -> Laberinto

Función que recibe una Trifurcacion, un Laberinto y un indicador de cual camino los relaciona y retorna una Trifurcacion en la que se indica que dicho camino especificado en el indicador conduce al Laberinto dado como parámetro. El indicador viene dado por un char i, r o d que indica si se agrega el Laberinto a la izquierda, recto o a la derecha respectivamente.

crearTesoro :: [Char] -> Laberinto -> Laberinto

Función que recibe un String con la descripción de un Tesoro y un Laberinto y retorna un Tesoro.

escribirLaberinto :: FilePath -> Laberinto -> IO ()

Funcion para escribir un laberinto

voltearIzquierda :: Laberinto -> Laberinto

Función que recibe un Laberinto y retorna el Laberinto que comienza al voltear al camino de la izquierda.

voltearDerecha :: Laberinto -> Laberinto

Función que recibe un Laberinto y retorna el Laberinto que comienza al voltear al camino de la derecha.

irRecto :: Laberinto -> Laberinto

Función que recibe un Laberinto y retorna el Laberinto que comienza al seguir por el camino recto.

recorrerLaberinto :: Laberinto -> [Char] -> Laberinto

Función que recibe un Laberinto y String que contiene una ruta y retorna el Laberinto que comienza en el punto al que conduce esa ruta. La ruta se especifica con indicadores que vienen dados por char i, r o d si es izquierda, recto o derecha respectivamente (por ejemplo dird representa el camino derecha-izquierda-recto-derecha).