Main

Documentation

main :: IO ()

Main

laberintoEnMem :: Laberinto

Función que crea una instancia de un Laberinto vacío en memoria.

opcionesDisponibles :: [(String, String)]

Función que define las opciones del menú.

printMenu :: IO ()

Función para mostrar el menú en pantalla.

instruccionesRuta :: IO ()

Función para mostrar en pantalla las instrucciones de cómo recibir una ruta para un Laberinto.

comenzarLaberintoNuevo :: IO ()

Función para construir un nuevo Laberinto a partir de una ruta dada por el usuario.

construirLaberinto :: Laberinto -> [Char] -> Laberinto

mostrarRecibirOpciones :: Laberinto -> IO ()

Funcion para mostrar las opciones y recibir del user